peopleproducer/requirements.txt
kafka-python==2.0.2
fastapi[all]==0.110.0
python-dotenv==1.0.1
Faker